pyproject.toml
[tool.poetry]
name = "applications_container"
version = "0.0.0"
description = ""
authors = ["Christoph Minixhofer <christoph.minixhofer@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
connexion = {extras = ["swagger-ui"], version = "^2.7.0"}
Flask-Cors = "^3.0.9"
waitress = "^1.4.4"
coverage = "^5.3"
bump2version = "^1.0.1"
toml = "^0.10.2"
coverage-badge = "^1.0.1"
google-auth = "^1.24.0"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
click = "^7.1.2"
pytest-env = "^0.6.2"

[tool.poetry.scripts]
container = 'applications_container:applications_container.main'
rename = 'applications_container:_rename._rename'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
